Refactored some code, mostyly in backend.

+++ backend/electron-ecommerce-rest/src/main/java/com/electron/rest/security/auth_service/RefreshTokenServiceImpl.java
@@ -1,6 +1,7 @@
 package com.electron.rest.security.auth_service;
 
 import com.electron.rest.exception.RefreshTokenException;
+import com.electron.rest.exception.UnauthorizedException;
 import com.electron.rest.security.AuthUtils;
 import com.electron.rest.security.auth_dto.LoginDto;
 import com.electron.rest.security.auth_entity.RefreshToken;
@@ -15,6 +16,8 @@ import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.ResponseCookie;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 
@@ -44,56 +47,51 @@ public class RefreshTokenServiceImpl implements RefreshTokenService {
     }
 
     @Override
-    public ResponseCookie getRefreshTokenCookie(LoginDto loginDto) {
+    public ResponseCookie getRefreshTokenCookie(LoginDto loginDto, Boolean remember) {
         List<UserProjection> usersList = userRepository.findUserIdFromEmail(loginDto.email());
         if (usersList.isEmpty())
             throw new UsernameNotFoundException("Bad credentials");
         Long userId = usersList.get(0).getId();
-        RefreshToken newToken = refreshTokenProvider.generateToken(userId);
-        refreshTokenRepository.save(newToken);
-        return refreshTokenProvider.createCookie(newToken.getToken());
+        if (remember) {
+            RefreshToken newToken = refreshTokenProvider.generateToken(userId);
+            refreshTokenRepository.save(newToken);
+            return refreshTokenProvider.createCookie(newToken.getToken());
+        }
+        return refreshTokenProvider.createClearCookie();
     }
 
     @Override
-    public ResponseCookie getLogoutCookie(HttpServletRequest request) {
-        String token = refreshTokenProvider.getTokenFromHttpRequest(request);
-
-        if (token == null) throw new RefreshTokenException(INVALID_TOKEN);
-        refreshTokenRepository.deleteRefreshTokenByToken(token);
+    public ResponseCookie getLogoutCookie(String refreshToken) {
+        if (refreshToken == null) throw new RefreshTokenException(INVALID_TOKEN);
+        refreshTokenRepository.deleteRefreshTokenByToken(refreshToken);
 
         return refreshTokenProvider.createClearCookie();
     }
 
     @Override
-    public Boolean logoutEverywhere(HttpServletRequest request) {
-        String token = AuthUtils.getTokenFromRequest(request);
-        String email = jwtProvider.getEmail(token);
+    public void logoutEverywhere(String jwt) {
+        if(jwt == null) throw new UnauthorizedException(INVALID_TOKEN);
+        String email = jwtProvider.getEmail(jwt);
+
         List<UserProjection> userList = userRepository.findUserIdFromEmail(email);
-        if(userList == null) throw new UsernameNotFoundException(USER_NOT_FOUND);
+        if (userList == null) throw new UsernameNotFoundException(USER_NOT_FOUND);
         Long userId = userList.get(0).getId();
         refreshTokenRepository.deleteRefreshTokenByUserId(userId);
-        return true;
     }
 
     @Override
-    public String isTokenUpToDate(HttpServletRequest request) {
+    public Boolean isTokenUpToDate(String token) {
+        if (token == null) throw new RefreshTokenException(INVALID_TOKEN);
+        List<RefreshTokenProjection> refreshTokenProjectionList = refreshTokenRepository.findRefreshTokenByToken(token);
+        if (refreshTokenProjectionList.isEmpty())
+            throw new RefreshTokenException(INVALID_TOKEN);
+        RefreshTokenProjection refreshTokenProjection = refreshTokenProjectionList.get(0);
 
-        String token = refreshTokenProvider.getTokenFromHttpRequest(request);
-        if (token != null) {
-
-            List<RefreshTokenProjection> refreshTokenProjectionList = refreshTokenRepository.findRefreshTokenByToken(token);
-            if (refreshTokenProjectionList.isEmpty())
-                throw new RefreshTokenException("Invalid token");
-            RefreshTokenProjection refreshTokenProjection = refreshTokenProjectionList.get(0);
-
-            if (refreshTokenProjection.getExpirationDate().compareTo(Instant.now()) < 0) {
-                refreshTokenRepository.deleteById(refreshTokenProjection.getId());
-                throw new RefreshTokenException("Refresh token expired. Please login.");
-            }
-            return token;
+        if (refreshTokenProjection.getExpirationDate().compareTo(Instant.now()) < 0) {
+            refreshTokenRepository.deleteById(refreshTokenProjection.getId());
+            throw new RefreshTokenException(EXPIRED_TOKEN);
         }
-        throw new RefreshTokenException(EXPIRED_TOKEN);
+        return true;
     }
-
 
 }
